streamlit/requirements.txt
altair
matplotlib
numpy
pandas
Pillow
scipy
streamlit
Prophet
sklearn
